#109 Adding application config
