+++ web/src/main/java/com/josdem/jmailer/JmailerApplication.java
@@ -1,60 +1,26 @@
 /*
-Copyright 2025 Jose Morales contact@josdem.io
+ Copyright 2025 Jose Morales contact@josdem.io
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
 
-http://www.apache.org/licenses/LICENSE-2.0
+   http://www.apache.org/licenses/LICENSE-2.0
 
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
- */
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
 
 package com.josdem.jmailer;
 
-import jakarta.jms.ConnectionFactory;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.jms.config.JmsListenerContainerFactory;
-import org.springframework.jms.config.SimpleJmsListenerContainerFactory;
-import org.springframework.mail.javamail.JavaMailSenderImpl;
 
 @SpringBootApplication
 public class JmailerApplication {
-
-  @Value("${email.username}")
-  private String username;
-
-  @Value("${email.password}")
-  private String password;
-
-  @Bean
-  public JmsListenerContainerFactory<?> myJmsContainerFactory(ConnectionFactory connectionFactory) {
-    var factory = new SimpleJmsListenerContainerFactory();
-    factory.setConnectionFactory(connectionFactory);
-    return factory;
-  }
-
-  @Bean
-  JavaMailSenderImpl javaMailSenderImpl() {
-    var mailSender = new JavaMailSenderImpl();
-    mailSender.setHost("smtp.gmail.com");
-    mailSender.setPort(587);
-    mailSender.setUsername(username);
-    mailSender.setPassword(password);
-    var prop = mailSender.getJavaMailProperties();
-    prop.put("mail.transport.protocol", "smtp");
-    prop.put("mail.smtp.auth", "true");
-    prop.put("mail.smtp.starttls.enable", "true");
-    prop.put("mail.debug", "true");
-    return mailSender;
-  }
 
   public static void main(String[] args) {
     SpringApplication.run(JmailerApplication.class, args);
